refactor : Change MenuItem inner class
