+++ app/src/main/java/com/bitcamp/goodplace/App.java
@@ -20,6 +20,7 @@ import com.bitcamp.goodplace.handler.BookmarkAddHandler;
 import com.bitcamp.goodplace.handler.BookmarkDeleteHandler;
 import com.bitcamp.goodplace.handler.BookmarkListHandler;
 import com.bitcamp.goodplace.handler.Command;
+import com.bitcamp.goodplace.handler.CommandRequest;
 import com.bitcamp.goodplace.handler.FullThemeListHandler;
 import com.bitcamp.goodplace.handler.MyMapAddHandler;
 import com.bitcamp.goodplace.handler.MyMapDeleteHandler;
@@ -43,189 +44,214 @@ import com.bitcamp.goodplace.handler.UserRankHandler;
 import com.bitcamp.goodplace.handler.UserUpdateHandler;
 import com.bitcamp.menu.Menu;
 import com.bitcamp.menu.MenuGroup;
-import com.bitcamp.menu.MenuItem;
 import com.bitcamp.util.Prompt;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
 public class App {
-  List<User> userList = new ArrayList<>();
-  HashMap<String, Command> commandMap = new HashMap<>();
+	List<User> userList = new ArrayList<>();
+	HashMap<String, Command> commandMap = new HashMap<>();
 
-  public App() {
-    commandMap.put("/auth/login", new AuthLoginHandler(userList));
-    commandMap.put("/auth/logout", new AuthLogoutHandler());
-    commandMap.put("/auth/displayLoginUer", new AuthDisplayLoginUserHandler());
+	class MenuItem extends Menu {
+		String menuId;
 
-    commandMap.put("/user/add", new UserAddHandler(userList));
-    commandMap.put("/user/delete", new UserDeleteHandler(userList));
-    commandMap.put("/user/detail", new UserDetailHandler(userList));
-    commandMap.put("/user/list", new UserListHandler(userList));
-    commandMap.put("/user/update", new UserUpdateHandler(userList));
+		public MenuItem(String title, String menuId) {
+			super(title);
+			this.menuId = menuId;
+		}
 
-    commandMap.put("/myMap/add", new MyMapAddHandler());
-    commandMap.put("/myMap/delete", new MyMapDeleteHandler());
-    commandMap.put("/myMap/list", new MyMapListHandler());
-    commandMap.put("/myMap/detail", new MyMapDetailHandler());
-    commandMap.put("/myMap/update", new MyMapUpdateHandler());
+		public MenuItem(String title, int enableState, String menuId) {
+			super(title, enableState);
+			this.menuId = menuId;
+		}
 
-    commandMap.put("/bookmark/add", new BookmarkAddHandler());
-    commandMap.put("/bookmark/delete", new BookmarkDeleteHandler());
-    commandMap.put("/bookmark/list", new BookmarkListHandler());
+		@Override
+		public void execute() {
+			Command command = commandMap.get(menuId);
+			try {
+				command.execute(new CommandRequest(commandMap));
+			} catch (Exception e) {
+				System.out.printf("%s 명령을 실행하는 중 오류 발생!", menuId);
+				e.printStackTrace();
+			}
+		}
 
-    commandMap.put("/place/add", new PlaceAddHandler());
-    commandMap.put("/place/delete", new PlaceDeleteHandler());
-    commandMap.put("/place/list", new PlaceListHandler());
+	}
 
-    commandMap.put("/search/list", new FullThemeListHandler(userList));
-    commandMap.put("/search/searchTheme", new SearchThemeHandler(userList));
-    commandMap.put("/search/searchUser", new SearchUserHandler(userList));
-    commandMap.put("/search/searchHashtag", new SearchHashtagHandler(userList));
+	public App() {
+		commandMap.put("/auth/login", new AuthLoginHandler(userList));
+		commandMap.put("/auth/logout", new AuthLogoutHandler());
+		commandMap.put("/auth/displayLoginUer", new AuthDisplayLoginUserHandler());
 
-    commandMap.put("/following/add", new UserFollowAddHandler(userList));
-    commandMap.put("/following/list", new UserFollowersListHandler(userList));
+		commandMap.put("/user/add", new UserAddHandler(userList));
+		commandMap.put("/user/delete", new UserDeleteHandler(userList));
+		commandMap.put("/user/detail", new UserDetailHandler(userList));
+		commandMap.put("/user/list", new UserListHandler(userList));
+		commandMap.put("/user/update", new UserUpdateHandler(userList));
 
-    commandMap.put("/rank/themeRank", new RealTimeRankHandler(userList));
-    commandMap.put("/rank/userRank", new UserRankHandler(userList));
-  }
+		commandMap.put("/myMap/add", new MyMapAddHandler());
+		commandMap.put("/myMap/delete", new MyMapDeleteHandler());
+		commandMap.put("/myMap/list", new MyMapListHandler());
+		commandMap.put("/myMap/detail", new MyMapDetailHandler());
+		commandMap.put("/myMap/update", new MyMapUpdateHandler());
 
-  public static void main(String[] args) {
-    App app = new App();
-    app.service();
-  }
+		commandMap.put("/bookmark/add", new BookmarkAddHandler());
+		commandMap.put("/bookmark/delete", new BookmarkDeleteHandler());
+		commandMap.put("/bookmark/list", new BookmarkListHandler());
 
-  void service() {
-    loadObject("user.json",userList,User.class);
+		commandMap.put("/place/add", new PlaceAddHandler());
+		commandMap.put("/place/delete", new PlaceDeleteHandler());
+		commandMap.put("/place/list", new PlaceListHandler());
 
-    createMenu().execute(commandMap);
-    Prompt.close();
+		commandMap.put("/search/list", new FullThemeListHandler(userList));
+		commandMap.put("/search/searchTheme", new SearchThemeHandler(userList));
+		commandMap.put("/search/searchUser", new SearchUserHandler(userList));
+		commandMap.put("/search/searchHashtag", new SearchHashtagHandler(userList));
 
-    saveObjects("user.json",userList);
-  }
+		commandMap.put("/following/add", new UserFollowAddHandler(userList));
+		commandMap.put("/following/list", new UserFollowersListHandler(userList));
 
-  public <E> void loadObject(String fileName, List<E> list, Class<E> domainType) {
-    try (BufferedReader in = new BufferedReader(new FileReader(fileName, Charset.forName("UTF-8")))) {
+		commandMap.put("/rank/themeRank", new RealTimeRankHandler(userList));
+		commandMap.put("/rank/userRank", new UserRankHandler(userList));
+	}
 
-      StringBuilder strBuilder = new StringBuilder();
-      String str;
-      while ((str = in.readLine()) != null) {
-        strBuilder.append(str);
-      }
-      Type type = TypeToken.getParameterized(Collection.class, domainType).getType();
-      Collection<E> collection = new Gson().fromJson(strBuilder.toString(), type);
+	public static void main(String[] args) {
+		App app = new App();
+		app.service();
+	}
 
-      list.addAll(collection);
+	void service() {
+		loadObject("user.json", userList, User.class);
 
-      System.out.printf("%s 데이터 로딩 완료!\n", fileName);
+		createMenu().execute();
+		Prompt.close();
 
-    } catch (Exception e) {
-      System.out.printf("%s 데이터 로딩 오류!\n", fileName);
-    }
-  }
+		saveObjects("user.json", userList);
+	}
 
-  private void saveObjects(String filepath, List<?> list) {
-    try (PrintWriter out = new PrintWriter(
-        new BufferedWriter(new FileWriter(filepath, Charset.forName("UTF-8"))))) {
+	public <E> void loadObject(String fileName, List<E> list, Class<E> domainType) {
+		try (BufferedReader in = new BufferedReader(new FileReader(fileName, Charset.forName("UTF-8")))) {
 
-      out.print(new Gson().toJson(list));
+			StringBuilder strBuilder = new StringBuilder();
+			String str;
+			while ((str = in.readLine()) != null) {
+				strBuilder.append(str);
+			}
+			Type type = TypeToken.getParameterized(Collection.class, domainType).getType();
+			Collection<E> collection = new Gson().fromJson(strBuilder.toString(), type);
 
-      System.out.printf("%s 데이터 출력 완료!\n", filepath);
+			list.addAll(collection);
 
-    } catch (Exception e) {
-      System.out.printf("%s 데이터 출력 오류!\n", filepath);
-      e.printStackTrace();
-    }
-  }
+			System.out.printf("%s 데이터 로딩 완료!\n", fileName);
 
-  Menu createMenu() {
-    MenuGroup mg = new MenuGroup("메인 메뉴");
-    mg.setPrevMenuTitle("종료");
+		} catch (Exception e) {
+			System.out.printf("%s 데이터 로딩 오류!\n", fileName);
+		}
+	}
 
-    mg.add(new MenuItem("로그인", Menu.ACCESS_LOGOUT, "/auth/login"));
-    mg.add(new MenuItem("내 정보", Menu.ACCESS_GENERAL, "/auth/displayLoginUer"));
-    mg.add(new MenuItem("로그아웃", Menu.ACCESS_GENERAL, "/auth/logout"));
+	private void saveObjects(String filepath, List<?> list) {
+		try (PrintWriter out = new PrintWriter(
+				new BufferedWriter(new FileWriter(filepath, Charset.forName("UTF-8"))))) {
 
-    createUserMenu(mg);
-    createMyMapMenu(mg);
-    //    createPlaceMenu(mg);
-    createsearchMenu(mg);
-    createBookmarkMenu(mg);
-    createRankMenu(mg);
-    createFollowMenu(mg);
+			out.print(new Gson().toJson(list));
 
-    return mg;
-  }
+			System.out.printf("%s 데이터 출력 완료!\n", filepath);
 
-  private Menu createUserMenu(MenuGroup mg) {
-    MenuGroup user = new MenuGroup("회원(회원가입)", Menu.ACCESS_LOGOUT | Menu.ACCESS_ADMIN);
-    user.add(new MenuItem("회원가입", Menu.ACCESS_LOGOUT, "/user/add"));
-    user.add(new MenuItem("회원목록", Menu.ACCESS_ADMIN, "/user/list"));
-    user.add(new MenuItem("회원상세", Menu.ACCESS_ADMIN, "/user/detail"));
-    user.add(new MenuItem("회원변경", Menu.ACCESS_ADMIN, "/user/update"));
-    user.add(new MenuItem("회원삭제", Menu.ACCESS_ADMIN, "/user/delete"));
+		} catch (Exception e) {
+			System.out.printf("%s 데이터 출력 오류!\n", filepath);
+			e.printStackTrace();
+		}
+	}
 
-    mg.add(user);
-    return user;
-  }
+	Menu createMenu() {
+		MenuGroup mg = new MenuGroup("메인 메뉴");
+		mg.setPrevMenuTitle("종료");
 
-  private void createMyMapMenu(MenuGroup mg) {
-    MenuGroup myMap = new MenuGroup("나만의 지도", Menu.ACCESS_ADMIN | Menu.ACCESS_GENERAL);
+		mg.add(new MenuItem("로그인", Menu.ACCESS_LOGOUT, "/auth/login"));
+		mg.add(new MenuItem("내 정보", Menu.ACCESS_GENERAL, "/auth/displayLoginUer"));
+		mg.add(new MenuItem("로그아웃", Menu.ACCESS_GENERAL, "/auth/logout"));
 
-    myMap.add(new MenuItem("테마 등록", "/myMap/add"));
-    myMap.add(new MenuItem("테마 목록", "/myMap/list"));
-    myMap.add(new MenuItem("테마 상세보기", "/myMap/detail"));
+		createUserMenu(mg);
+		createMyMapMenu(mg);
+		// createPlaceMenu(mg);
+		createsearchMenu(mg);
+		createBookmarkMenu(mg);
+		createRankMenu(mg);
+		createFollowMenu(mg);
+
+		return mg;
+	}
+
+	private Menu createUserMenu(MenuGroup mg) {
+		MenuGroup user = new MenuGroup("회원(회원가입)", Menu.ACCESS_LOGOUT | Menu.ACCESS_ADMIN);
+		user.add(new MenuItem("회원가입", Menu.ACCESS_LOGOUT, "/user/add"));
+		user.add(new MenuItem("회원목록", Menu.ACCESS_ADMIN, "/user/list"));
+		user.add(new MenuItem("회원상세", Menu.ACCESS_ADMIN, "/user/detail"));
+		user.add(new MenuItem("회원변경", Menu.ACCESS_ADMIN, "/user/update"));
+		user.add(new MenuItem("회원삭제", Menu.ACCESS_ADMIN, "/user/delete"));
+
+		mg.add(user);
+		return user;
+	}
+
+	private void createMyMapMenu(MenuGroup mg) {
+		MenuGroup myMap = new MenuGroup("나만의 지도", Menu.ACCESS_ADMIN | Menu.ACCESS_GENERAL);
+
+		myMap.add(new MenuItem("테마 등록", "/myMap/add"));
+		myMap.add(new MenuItem("테마 목록", "/myMap/list"));
+		myMap.add(new MenuItem("테마 상세보기", "/myMap/detail"));
 //    myMap.add(new MenuItem("테마 수정", "/myMap/update"));
-    myMap.add(new MenuItem("테마 삭제", "/myMap/delete"));
+		myMap.add(new MenuItem("테마 삭제", "/myMap/delete"));
 
-    mg.add(myMap);
-  }
+		mg.add(myMap);
+	}
 
-  //  private void createPlaceMenu(MenuGroup mg) {
-  //    MenuGroup savePlaceInTheme = new MenuGroup("테마에 장소 추가", Menu.ACCESS_ADMIN | Menu.ACCESS_GENERAL);
-  //
-  //    savePlaceInTheme.add(new MenuItem("장소 등록", "/place/add"));
-  //    savePlaceInTheme.add(new MenuItem("장소 목록", "/place/list"));
-  //    savePlaceInTheme.add(new MenuItem("장소 삭제", "/place/delete"));
-  //
-  //    mg.add(savePlaceInTheme);
-  //  }
+	// private void createPlaceMenu(MenuGroup mg) {
+	// MenuGroup savePlaceInTheme = new MenuGroup("테마에 장소 추가", Menu.ACCESS_ADMIN |
+	// Menu.ACCESS_GENERAL);
+	//
+	// savePlaceInTheme.add(new MenuItem("장소 등록", "/place/add"));
+	// savePlaceInTheme.add(new MenuItem("장소 목록", "/place/list"));
+	// savePlaceInTheme.add(new MenuItem("장소 삭제", "/place/delete"));
+	//
+	// mg.add(savePlaceInTheme);
+	// }
 
-  private void createsearchMenu(MenuGroup mg) {
-    MenuGroup search = new MenuGroup("검색");
+	private void createsearchMenu(MenuGroup mg) {
+		MenuGroup search = new MenuGroup("검색");
 
-    search.add(new MenuItem("전체 테마 목록", "/search/list"));
-    search.add(new MenuItem("테마 검색", "/search/searchTheme"));
-    search.add(new MenuItem("해시태그 검색", "/search/searchHashtag"));
-    search.add(new MenuItem("유저 검색", "/search/searchUser"));
+		search.add(new MenuItem("전체 테마 목록", "/search/list"));
+		search.add(new MenuItem("테마 검색", "/search/searchTheme"));
+		search.add(new MenuItem("해시태그 검색", "/search/searchHashtag"));
+		search.add(new MenuItem("유저 검색", "/search/searchUser"));
 
+		mg.add(search);
+	}
 
-    mg.add(search);
-  }
+	private void createBookmarkMenu(MenuGroup mg) {
+		MenuGroup bookmark = new MenuGroup("북마크", Menu.ACCESS_ADMIN | Menu.ACCESS_GENERAL);
 
-  private void createBookmarkMenu(MenuGroup mg) {
-    MenuGroup bookmark = new MenuGroup("북마크", Menu.ACCESS_ADMIN | Menu.ACCESS_GENERAL);
+		bookmark.add(new MenuItem("북마크 등록", "/bookmark/add"));
+		bookmark.add(new MenuItem("북마크 목록", "/bookmark/list"));
+		bookmark.add(new MenuItem("북마크 삭제", "/bookmark/delete"));
 
-    bookmark.add(new MenuItem("북마크 등록", "/bookmark/add"));
-    bookmark.add(new MenuItem("북마크 목록", "/bookmark/list"));
-    bookmark.add(new MenuItem("북마크 삭제", "/bookmark/delete"));
+		mg.add(bookmark);
+	}
 
-    mg.add(bookmark);
-  }
+	private void createRankMenu(MenuGroup mg) {
+		MenuGroup rank = new MenuGroup("순위");
 
-  private void createRankMenu(MenuGroup mg) {
-    MenuGroup rank = new MenuGroup("순위");
+		rank.add(new MenuItem("테마 순위", "/rank/themeRank"));
+		rank.add(new MenuItem("유저 순위", "/rank/userRank"));
 
-    rank.add(new MenuItem("테마 순위", "/rank/themeRank"));
-    rank.add(new MenuItem("유저 순위", "/rank/userRank"));
+		mg.add(rank);
+	}
 
-    mg.add(rank);
-  }
+	private void createFollowMenu(MenuGroup mg) {
+		MenuGroup follow = new MenuGroup("팔로우");
+		follow.add(new MenuItem("팔로우 추가", "/following/add"));
+		follow.add(new MenuItem("팔로잉 목록 ", "/following/list"));
 
-  private void createFollowMenu(MenuGroup mg) {
-    MenuGroup follow = new MenuGroup("팔로우");
-    follow.add(new MenuItem("팔로우 추가", "/following/add"));
-    follow.add(new MenuItem("팔로잉 목록 ", "/following/list"));
-
-    mg.add(follow);
-  }
+		mg.add(follow);
+	}
 }
